Struct lattice_evolution::lattice::Lattice

pub struct Lattice {
    g: f64,
    q: f64,
    depth: f64,
    time: f64,
    h0: DMatrix<Complex64>,
    h1: DMatrix<Complex64>,
    h2: DMatrix<Complex64>,
    psi: DVector<Complex64>,
    dpsi_a: DVector<Complex64>,
    dpsi_v: DVector<Complex64>,
}

Lattice struct represents an instance of the Shaken Optical Lattice

Fields
g: f64

acceleration/gravity

q: f64

Correction to pure momentum, $(p+q)$ is kinematic momentum

depth: f64

lattice depth

time: f64

time

h0: DMatrix<Complex64>

The Kinetic part of the Hamiltonian

h1: DMatrix<Complex64>

The $sin(\phi)$ part of the lattice coupling

h2: DMatrix<Complex64>

the $\cos(\phi)$ part of the lattice coupling

psi: DVector<Complex64>

current wavefunction of wavepacket moving through lattice

dpsi_a: DVector<Complex64>dpsi_v: DVector<Complex64>
Implementations

impl Lattice

pub fn get_acceleration(&self) -> f64

Returns $a$ of this Lattice

pub fn get_depth(&self) -> f64

Returns $V_0$ of this Lattice

pub fn set_depth(&mut self, depth: f64)

Sets $V_0$ of this Lattice

pub fn set_time(&mut self, time: f64)

Sets time of this lattice. Used only when we’re manually doing RK4 to track CFI and QFI metrics.

pub fn get_hamiltonian(&self) -> DMatrix<Complex64>

Returns full Hamiltonian with $\phi=0$

pub fn get_psi(&self) -> DVector<Complex64>

Current Wavefunction of the wavepacket

pub fn get_time(&self) -> f64

pub fn get_dpsi_a(&self) -> DVector<Complex64>

pub fn get_momentum(&self) -> DVector<f64>

pub fn get_dmomentum_da(&self) -> DVector<f64>

pub fn accelerate(&mut self, impulse: f64)

Accelerate the wavepacket by adding an impulse to $q$ and updating kinematic momentum

pub fn step(&mut self, amplitude: f64, omega: f64)

Evolve wavepacket in this lattice given $(A,\omega)$ for $\phi=A\sin(\omega t)$ for a time period $period = \pi/(11.5 \omega)$. Usually, we set $\omega=11.5\omega_r$ to be constant, given by FREQ in Lattice module

the step() method calls the rk4step() solver, which uses Runge-Kutta 4 to update the system. To do so it calls update(), the method for calculating derivatives

pub fn update( &self, wavefunction: DVector<Complex64>, amplitude: f64, omega: f64, t: f64 ) -> DVector<Complex64>

update is essentially the derivative to the wavepacket at time t$ It’s given by $-iH|\psi>$

pub fn update_da( &self, state_deriv: DVector<Complex64>, state: DVector<Complex64>, amplitude: f64, omega: f64, t: f64 ) -> DVector<Complex64>

update_d is essentially the derivative to the state derivative at time t$ It’s given by $-iH|\psi> +i*\hat{p}t$

pub fn update_d_v( &self, state_deriv: DVector<Complex64>, state: DVector<Complex64>, amplitude: f64, omega: f64, t: f64 ) -> DVector<Complex64>

pub fn rk4step(&mut self, dt: f64, amplitude: f64, omega: f64, t: f64)

The RK4 step.

impl Lattice

pub fn new(acceleration: f64, latticedepth: f64) -> Lattice

Initializes the Lattice and makes a new instnce Uses given values of $(a,V_0)$ (see paper) to construct the Hamiltonian. It further initializes the wavepacket to be the ground state.

By default, the const N_STATES defines how many momentum states are included in the simulation basis. So N_STATES=11 means $p\in {-10\hbark_L, -8\hbar k_L, …, 10\hbar k_L}$

impl Lattice

pub fn acc_qfi(&self) -> f64

pub fn acc_cfi(&self) -> f64

pub fn depth_qfi(&self) -> f64

pub fn depth_cfi(&self) -> f64

pub fn acc_depth_cfi(&self) -> f64