Struct lattice_evolution::lattice::Lattice

pub struct Lattice {
    g: f64,
    q: f64,
    depth: f64,
    h0: DMatrix<Complex64>,
    h1: DMatrix<Complex64>,
    h2: DMatrix<Complex64>,
    psi: DVector<Complex64>,
}

Lattice struct represents an instance of the Shaken Optical Lattice

Fields
g: f64q: f64depth: f64h0: DMatrix<Complex64>h1: DMatrix<Complex64>h2: DMatrix<Complex64>psi: DVector<Complex64>
Implementations

impl Lattice

pub fn get_depth(&self) -> f64

Returns $V_0$ of this Lattice

pub fn set_depth(&mut self, depth: f64)

Sets $V_0$ of this Lattice

pub fn get_hamiltonian(&self) -> DMatrix<Complex64>

Returns full Hamiltonian with $\phi=0$

pub fn get_psi(&self) -> DVector<Complex64>

Current Wavefunction of the wavepacket

pub fn accelerate(&mut self, impulse: f64)

Accelerate the wavepacket by adding an impulse to $q$ and updating kinematic momentum

pub fn step(&mut self, amplitude: f64, omega: f64)

Evolve wavepacket in this lattice given $(A,\omega)$ for $\phi=A\sin(\omega t)$ for a time period $period = \pi/(11.5 \omega)$. Usually, we set $\omega=11.5\omega_r$ to be constant, given by FREQ in Lattice module

the step() method calls the rk4step() solver, which uses Runge-Kutta 4 to update the system. To do so it calls update(), the method for calculating derivatives

pub fn update( &self, wavefunction: DVector<Complex64>, amplitude: f64, omega: f64, t: f64 ) -> DVector<Complex64>

update is essentially the derivative to the wavepacket at time t$ It’s given by $-iH|\psi>$

pub fn rk4step(&mut self, dt: f64, amplitude: f64, omega: f64, t: f64)

The RK4 step.

impl Lattice

pub fn new(acceleration: f64, latticedepth: f64) -> Lattice

Initializes the Lattice and makes a new instnce Uses given values of $(a,V_0)$ (see paper) to construct the Hamiltonian. It further initializes the wavepacket to be the ground state.

By default, the const N_STATES defines how many momentum states are included in the simulation basis. So N_STATES=11 means $p\in {-10\hbark_L, -8\hbar k_L, …, 10\hbar k_L}$